+++ tasks/seq/rezantseva_a_vector_dot_product/src/ops_seq.cpp
@@ -1,19 +1,18 @@
 // Copyright 2024 Nesterov Alexander
-#include <random>
 #include "seq/rezantseva_a_vector_dot_product/include/ops_seq.hpp"
-
+#include <random>
 #include <thread>
 
-static int offset = 0; 
+static int offset = 0;
 using namespace std::chrono_literals;
 
 bool rezantseva_a_vector_dot_product_seq::TestTaskSequential::validation() {
   internal_order_test();
   // Check count elements of output
   return (taskData->inputs.size() == taskData->inputs_count.size() && taskData->inputs.size() == 2) &&
-         (taskData->inputs_count[0] == taskData->inputs_count[1]) && (taskData->outputs.size() == taskData->outputs_count.size()) && 
-         taskData->outputs.size() == 1 && taskData->outputs_count[0] == 1;
-
+         (taskData->inputs_count[0] == taskData->inputs_count[1]) &&
+         (taskData->outputs.size() == taskData->outputs_count.size()) && taskData->outputs.size() == 1 &&
+         taskData->outputs_count[0] == 1;
 }
 
 bool rezantseva_a_vector_dot_product_seq::TestTaskSequential::pre_processing() {
@@ -21,12 +20,10 @@ bool rezantseva_a_vector_dot_product_seq::TestTaskSequential::pre_processing() {
   // Init value for input and output
 
   input_ = std::vector<std::vector<int>>(taskData->inputs.size());
-  for (size_t i = 0; i < input_.size(); i++) 
-  {
+  for (size_t i = 0; i < input_.size(); i++) {
     auto* tmp_ptr = reinterpret_cast<int*>(taskData->inputs[i]);
     input_[i] = std::vector<int>(taskData->inputs_count[i]);
-    for (size_t  j = 0; j < taskData->inputs_count[i]; j++) 
-    {
+    for (size_t  j = 0; j < taskData->inputs_count[i]; j++) {
       input_[i][j] = tmp_ptr[j];
     }
   }
@@ -36,8 +33,7 @@ bool rezantseva_a_vector_dot_product_seq::TestTaskSequential::pre_processing() {
 
 bool rezantseva_a_vector_dot_product_seq::TestTaskSequential::run() {
   internal_order_test();
-  for (size_t i = 0; i < input_[0].size(); i++) 
-  {
+  for (size_t i = 0; i < input_[0].size(); i++) {
     res += input_[0][i] * input_[1][i];
   }
 
@@ -54,14 +50,12 @@ std::vector<int> rezantseva_a_vector_dot_product_seq::createRandomVector(const i
   std::vector<int> vec(v_size);
   std::mt19937 gen;
   gen.seed((unsigned)time(0) + ++offset);
-  for (int i = 0; i < v_size; i++) 
-      vec[i] = gen() % 100;             
+  for (int i = 0; i < v_size; i++) vec[i] = gen() % 100;
   return vec;
 }
 int rezantseva_a_vector_dot_product_seq::vectorDotProduct(const std::vector<int>& v1, const std::vector<int>& v2) {
   if (v1.size() != v2.size()) throw "Error! Vectors must have equal size!";
   long long result = 0;
-  for (int i = 0; i < v1.size(); i++) 
-      result += v1[i] * v2[i];
+  for (size_t i = 0; i < v1.size(); i++) result += v1[i] * v2[i];
   return result;
 }
